prevent re-execution

+++ utils/split_depth_file.py
@@ -1,5 +1,6 @@
 import sys
 import re
+from os.path import exists
 
 from get_locus import get_locus
 
@@ -20,14 +21,15 @@ def split_depth_file(file_path,reference_gb):
     reference_list = get_locus(reference_gb)
     print("Reference list === ",reference_list)
     for name in reference_list:
-        with open(file_path, "r") as f:
-            new_file = []
-            for line in f.readlines():
-                print(line.split()[0] , str(name))
-                if line.split()[0] == str(name):
-                    new_file.append(line)
-            with open(f"{path}{name}.depth", "w") as output_file:
-                output_file.writelines(new_file)   
+        if not exists(f"{path}{name}.depth"):
+            with open(file_path, "r") as f:
+                new_file = []
+                for line in f.readlines():
+                    print(line.split()[0] , str(name))
+                    if line.split()[0] == str(name):
+                        new_file.append(line)
+                    with open(f"{path}{name}.depth", "w") as output_file:
+                        output_file.writelines(new_file)   
 
 
 if __name__ == '__main__':
